Move common APIs from rabbit `module` to `core`
Some exceptions and listener supporting classes
are not only for RabbitMQ integration.
They can be used in any other Spring AMQP implementations.

* Deprecate existing classes in favor of their new counterpart
with extensions where possible.
* Still use instances of deprecated classes for proper upper bound
of the consumers of the target API, essentially following Postel's Law
* Move `ContainerUtilsTests` to `core` module as we don't test a deprecated class any more
* Add `reactor-core` optional dependency to `core` module
to satisfy `ContinuationHandlerMethodArgumentResolver` requirements
