+++ spring-amqp/src/main/java/org/springframework/amqp/listener/FatalExceptionStrategy.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2010-present the original author or authors.
+ * Copyright 2026-present the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,16 +14,22 @@
  * limitations under the License.
  */
 
-package org.springframework.amqp.rabbit.listener.exception;
+package org.springframework.amqp.listener;
+
 
 /**
- * Exception class that indicates a rejected message on shutdown. Used to trigger a rollback for an
- * external transaction manager in that case.
+ * A strategy interface for the {@code ConditionalRejectingErrorHandler} to
+ * decide whether an exception should be considered as fatal and the
+ * message should not be requeued (released), rather discarded (rejected).
  *
- * @deprecated in favor of {@link org.springframework.amqp.listener.MessageRejectedWhileStoppingException}.
+ * @author Gary Russell
+ * @author Artem Bilan
+ *
+ * @since 4.1
  */
-@Deprecated(forRemoval = true, since = "4.1")
-@SuppressWarnings("serial")
-public class MessageRejectedWhileStoppingException extends org.springframework.amqp.listener.MessageRejectedWhileStoppingException {
+@FunctionalInterface
+public interface FatalExceptionStrategy {
+
+	boolean isFatal(Throwable throwable);
 
 }
