+++ spring-amqp/src/main/java/org/springframework/amqp/listener/adapter/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Package for Spring AMQP common listener adapter support classes.
+ */
+@org.jspecify.annotations.NullMarked
+package org.springframework.amqp.listener.adapter;
