+++ spring-amqp/src/main/java/org/springframework/amqp/listener/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Package for Spring AMQP common listener support classes.
+ */
+@org.jspecify.annotations.NullMarked
+package org.springframework.amqp.listener;
